warden-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
bcrypt==5.0.0
python-jose[cryptography]==3.3.0
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
pyyaml==6.0.3
